Moved migration into it's own file
