+++ packages/grid_client/src/helpers/index.ts
@@ -3,3 +3,4 @@ export * from "./requests";
 export * from "./events";
 export * from "./validator";
 export * from "./expose";
+export * from "./migration";
